Implement generic tabs page
